requirements.txt
asgiref==3.5.0
attrs==21.4.0
black==22.3.0
click==8.1.3
Django==3.2
django-debug-toolbar==3.2.4
django-filter==22.1
djangorestframework==3.13.1
drf-nested-routers==0.93.4
drf-spectacular==0.22.1
importlib-resources==5.8.0
inflection==0.5.1
jsonschema==4.6.0
mypy-extensions==0.4.3
mysqlclient==2.1.0
pathspec==0.9.0
platformdirs==2.5.2
pyrsistent==0.18.1
python-dotenv==0.20.0
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.2.0
uritemplate==4.1.1
zipp==3.8.0
